Fix Plug.BodyReader.Deflate when some output is queued by :zlib
To prevent zip bombs, :zlib.safeInflate/2 does not return all the output
at once if it is larger than a certain size. Instead, it queues the
ouput and returns it chunk by chunk on successive calls. We thus need to
drain the queue to make sure we have all the data before closing the
stream.